Add custom derive support.
This adds support for expanding custom derives at runtime. It does so
by allowing the user to provide a custom derive function at runtime
that takes a syn::DeriveInput and returns a proc_macro2::TokenStream.

The use of syn::DeriveInput as an input rather than a TokenStream
is due to the fact that the method to convert a TokenStream to a
DeriveInput can itself only be called in the context of a proc_macro,
and as such is inconvenient for end users to use in tests.

Because the type of the user provided function is different, it is
provided as a new function, emulate_derive_expansion_fallible. I did not
provide an equivalent for the deprecated emulate_macro_expansion function.

It works by visiting all structs and enums, and checking them for a
derive attribute that derives the user provided name for their custom
derive. This has the same limitations with regards to only checking the
final path component as the existing function. When it finds a match, it
converts the ItemStruct or ItemEnum to a DeriveInput and calls the user
provided function with it.

This required updating proc_macro2 to the newest version, which in turn required
updating syn to the newest version. Since this was a widespread change, I took
the opportunity to refactor to the rust 2018 edition in the process.

Let me know if you have any feedback with how this was implemented.